Removing getFieldsFromInput() from editService()
getFieldsFromInput essentially sets default values for input.
In editService, this doesn't work if we remove the password field.

If a field isn't provided, its key is still set and its value is null.

Since null isn't going to match what's in the database, this will
be added to $edit_fields and null will be sent as $post_vps['rootpass']
and set as the rootpass in the database.

If $vars is directly used instead of $params, this won't happen.